Fix issues with planet filtering and building.
