+++ src/data/GameData.ts
@@ -1,9 +1,9 @@
-// import stableHash from "stable-hash";
-// import useSWR from "swr";
-// import { BASE_OPTIONS } from "../../server/data/options";
-// import { fetcher } from "../util/api/util";
-// import { validateMapString } from "../util/util";
 import { IntlShape } from "react-intl";
+import {
+  updateMapString,
+  isValidMapString,
+  validSystemNumber,
+} from "../util/map";
 
 let getBaseAgendas: DataFunction<AgendaId, BaseAgenda> = () => {
   return {};
@@ -453,10 +453,20 @@ export function buildPlanets(storedGameData: StoredGameData) {
   const gameFactions = storedGameData.factions ?? {};
   const gameOptions = storedGameData.options;
 
+  const numFactions = Object.keys(gameFactions).length;
   const mapString = gameOptions["map-string"] ?? "";
-  const isValidMapString = validateMapString(mapString);
+  const mapStyle = gameOptions["map-style"] ?? "standard";
+  const updatedMapString = updateMapString(mapString, mapStyle, numFactions);
+  const validMapString = isValidMapString(updatedMapString, numFactions);
   const inGameSystems = mapString
     .split(" ")
+    .filter(validSystemNumber)
+    .filter(
+      (systemNumber) =>
+        systemNumber !== "-1" &&
+        systemNumber !== "0" &&
+        !isNaN(parseInt(systemNumber))
+    )
     .map((system) => parseInt(system) as SystemId);
 
   let planets = {} as Partial<Record<PlanetId, Planet>>;
@@ -479,7 +489,8 @@ export function buildPlanets(storedGameData: StoredGameData) {
       return;
     }
     if (
-      isValidMapString &&
+      validMapString &&
+      inGameSystems.length > 0 &&
       planet.system &&
       planet.id !== "Mirage" &&
       planet.id !== "Mallice" &&
